+++ src/components/Layout/index.tsx
@@ -1,7 +1,7 @@
 import { PropsWithChildren, useState } from 'react';
 import { Navbar } from '../Navbar';
 import { ContentWrapper, Wrapper } from './index.styles';
-import { Modal } from '@/components/Modal';
+import { Modal } from '@/modals/Modal';
 import { CreateCard } from '@/components/CreateCard';
 
 export default function Layout({ children }: PropsWithChildren) {
