+++ src/modals/Modal/index.styles.ts
